Rename: NotFound 컴포넌트 NetworkError로 네이밍 변경
