+++ frontend/src/components/common/error/index.ts
@@ -1,0 +1,2 @@
+export { NetworkError } from './NetworkError'
+export { NotFoundError } from './NotFoundError'
