+++ frontend/src/pages/admin/member/components/expel-member-table/ExpelMemberTable.tsx
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query'
 import { ColumnDef } from '@tanstack/react-table'
 
-import { NotFound } from '@/components/common'
+import { NetworkError } from '@/components/common'
 import {
   MemberTable,
   MemberTableSkeleton,
@@ -69,7 +69,7 @@ export const ExpelMemberTable = () => {
   if (status === 'pending')
     return <MemberTableSkeleton columns={expelMemberColumn} />
 
-  if (error) return <NotFound />
+  if (error) return <NetworkError />
 
   if (!activeUsers.length) {
     return (
